Move Promo to a job
